+++ hail/src/main/scala/is/hail/backend/HailTaskContext.scala
@@ -23,8 +23,7 @@ abstract class HailTaskContext {
   def finish(): Unit = {
     log.info(s"TaskReport: stage=${ stageId() }, partition=${ partitionId() }, attempt=${ attemptNumber() }, " +
       s"peakBytes=${ thePool.getHighestTotalUsage }, peakBytesReadable=${ formatSpace(thePool.getHighestTotalUsage) }"+
-      s"chunks requested=${thePool.getUsage._1}, cache hits=${thePool.getUsage._2}, " +
-      s"dealloc to fit =${thePool.getUsage._3}")
+      s"chunks requested=${thePool.getUsage._1}, cache hits=${thePool.getUsage._2}")
     thePool.close()
   }
 }
